% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{authorize}
\alias{authorize}
\title{Authorize app to use Danfoss Ally API}
\usage{
authorize(key, secret)
}
\arguments{
\item{key, }{secret character strings}
}
\description{
To obtain the required key and secret, register at https://developer.danfoss.com/
}
