% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{devices}
\alias{devices}
\title{Get device and status information}
\usage{
devices(id = NULL)
}
\value{
a single tibble combining device and status information
}
\description{
Get device and status information
}
