% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{get_data}
\alias{get_data}
\title{Get device information and status from Danfoss Ally API}
\usage{
get_data(id = NULL)
}
\arguments{
\item{id}{a string}
}
\value{
a list
}
\description{
Get device information and status from Danfoss Ally API
}
