% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_devices.R
\name{get_devices}
\alias{get_devices}
\title{Get device information and status from Danfoss Ally API}
\usage{
get_devices(id = NULL)
}
\arguments{
\item{id}{a string}
}
\value{
a list
}
\description{
Get device information and status from Danfoss Ally API
}
